refactoring: 코드 수정
