--- .github/PULL_REQUEST_TEMPLATE.md
@@ -1,7 +1,0 @@
-## 요약
-
-## 관련 이슈
-
-## 변경 사항
-
-## 참고 사항
