--- .prettierrc.js
@@ -1,7 +1,0 @@
-module.exports = {
-  semi: true,
-  trailingComma: 'es5',
-  singleQuote: true,
-  printWidth: 80,
-  endOfLine: 'auto',
-};
