--- README.md
@@ -1,1 +1,0 @@
-# JavaScript로 Todo App 만들기
