+++ index.html
@@ -1,19 +1,18 @@
-<!doctype html>
+<!Doctype html>
 <html lang="ko">
   <head>
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <!-- <link rel="icon" href="/favicon.ico" type="image/x-icon" /> -->
     <link rel="stylesheet" href="style.css" />
-    <link rel="stylesheet" href="todo.css" />
+    <link rel="stylesheet" href="src/styles/todo.css" />
     <title>Todo App</title>
   </head>
   <body>
     <h1 class="title">투두 리스트</h1>
-    <div class="container">
+    <div class="todoContainer">
       <div>
         <input
-          id="taskInput"
+          id="todoInput"
           class="todoInput"
           type="text"
           placeholder="할일을 추가해주세요!"
@@ -21,7 +20,7 @@
           autocomplete="off"
         />
       </div>
-      <ul id="taskList"></ul>
+      <ul id="taskList" class="taskList"></ul>
     </div>
     <script type="module" src="index.js"></script>
   </body>
