+++ index.js
@@ -1,8 +1,8 @@
-import { initializeTaskInput } from './taskInput.js';
-import { createTaskItemElement } from './taskItem.js';
+import { initializeTaskInput } from './src/taskInput.js';
+import { createTaskItemElement } from './src/taskItem.js';
 import { $ } from './utilFunc.js';
 
-const $taskInput = $('taskInput');
+const $taskInput = $('todoInput');
 const $taskList = $('taskList');
 
 initializeTaskInput($taskInput, $taskList, createTaskItemElement);
