--- package.json
@@ -1,27 +1,0 @@
-{
-  "devDependencies": {
-    "eslint": "^8.48.0",
-    "eslint-config-prettier": "^9.0.0",
-    "eslint-plugin-html": "^7.1.0",
-    "eslint-plugin-prettier": "^5.0.0",
-    "prettier": "^3.0.3"
-  },
-  "name": "js-todo-app",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/kzh4295/js-todo-app.git"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "bugs": {
-    "url": "https://github.com/kzh4295/js-todo-app/issues"
-  },
-  "homepage": "https://github.com/kzh4295/js-todo-app#readme"
-}
