+++ utilFunc.js
@@ -1,4 +1,3 @@
-// utils
 export function $(selector) {
   return document.getElementById(selector);
 }
